Addresses #606 - fixes inconsistant casts of use_graphics. It was created in server as a bool, then was used as a byte (use_graphics == 3), etc. Change all references to int, to match client side.
git-svn-id: http://developer.mangband.org/svn/mangband/trunk@1352 802d84f3-b497-254e-bede-06c0a5a3456a

+++ src/common/types.h
@@ -1347,7 +1347,7 @@ struct player_type
 	char tval_char[128];
    
 	byte dungeon_stream;
-	byte use_graphics;
+	int use_graphics;
 	byte screen_wid;
 	byte screen_hgt;
